requirements-dev.txt
cachetools
flit
gidgethub[aiohttp]
ipython
pytest
pytest-cov
pytest-aiohttp
